refactor: 가게 분류를 가게 하위로, 상품 분류를 상품 하위로 구조 변경 (#17)
