+++ src/main/java/com/baedalping/delivery/domain/product/productCategory/dto/ProductCategoryCreateRequestDto.java
@@ -1,4 +1,4 @@
-package com.baedalping.delivery.domain.productCategory.dto;
+package com.baedalping.delivery.domain.product.productCategory.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
@@ -8,7 +8,7 @@ import lombok.Setter;
 @Setter
 @Getter
 @NoArgsConstructor
-public class ProductCategoryUpdateRequestDto {
+public class ProductCategoryCreateRequestDto {
   @NotBlank(message = "상품 분류명을 입력해주세요.")
   private String productCategoryName;
 }
