+++ src/main/java/com/baedalping/delivery/domain/product/productCategory/service/ProductCategoryService.java
@@ -1,13 +1,13 @@
-package com.baedalping.delivery.domain.productCategory.service;
+package com.baedalping.delivery.domain.product.productCategory.service;
 
 
 
-import com.baedalping.delivery.domain.productCategory.dto.ProductCategoryCreateRequestDto;
-import com.baedalping.delivery.domain.productCategory.dto.ProductCategoryCreateResponseDto;
-import com.baedalping.delivery.domain.productCategory.dto.ProductCategoryUpdateRequestDto;
-import com.baedalping.delivery.domain.productCategory.dto.ProductCategoryUpdateResponseDto;
-import com.baedalping.delivery.domain.productCategory.entity.ProductCategory;
-import com.baedalping.delivery.domain.productCategory.repository.ProductCategoryRepository;
+import com.baedalping.delivery.domain.product.productCategory.dto.ProductCategoryCreateRequestDto;
+import com.baedalping.delivery.domain.product.productCategory.dto.ProductCategoryCreateResponseDto;
+import com.baedalping.delivery.domain.product.productCategory.dto.ProductCategoryUpdateRequestDto;
+import com.baedalping.delivery.domain.product.productCategory.dto.ProductCategoryUpdateResponseDto;
+import com.baedalping.delivery.domain.product.productCategory.entity.ProductCategory;
+import com.baedalping.delivery.domain.product.productCategory.repository.ProductCategoryRepository;
 import com.baedalping.delivery.global.common.exception.DeliveryApplicationException;
 import com.baedalping.delivery.global.common.exception.ErrorCode;
 import jakarta.transaction.Transactional;
