+++ src/main/java/com/baedalping/delivery/domain/store/service/StoreService.java
@@ -9,8 +9,8 @@ import com.baedalping.delivery.domain.store.entity.Store;
 import com.baedalping.delivery.domain.store.repository.StoreRepository;
 import com.baedalping.delivery.global.common.exception.DeliveryApplicationException;
 import com.baedalping.delivery.global.common.exception.ErrorCode;
-import com.baedalping.delivery.domain.storeCategory.entity.StoreCategory;
-import com.baedalping.delivery.domain.storeCategory.repository.StoreCategoryRepository;
+import com.baedalping.delivery.domain.store.storeCategory.entity.StoreCategory;
+import com.baedalping.delivery.domain.store.storeCategory.repository.StoreCategoryRepository;
 import jakarta.transaction.Transactional;
 import java.util.UUID;
 import lombok.AllArgsConstructor;
