+++ src/main/java/com/baedalping/delivery/domain/store/storeCategory/dto/StoreCategoryUpdateResponseDto.java
@@ -1,17 +1,17 @@
-package com.baedalping.delivery.domain.storeCategory.dto;
+package com.baedalping.delivery.domain.store.storeCategory.dto;
 
-import com.baedalping.delivery.domain.storeCategory.entity.StoreCategory;
+import com.baedalping.delivery.domain.store.storeCategory.entity.StoreCategory;
 import java.util.UUID;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class StoreCategoryCreateResponseDto {
+public class StoreCategoryUpdateResponseDto {
   private UUID storeCategoryId;
   private String storeCategoryName;
 
-  public StoreCategoryCreateResponseDto(StoreCategory storeCategory){
+  public StoreCategoryUpdateResponseDto(StoreCategory storeCategory){
     this.storeCategoryId = storeCategory.getStoreCategoryId();
     this.storeCategoryName = storeCategory.getStoreCategoryName();
   }
